[Refactor] Rearrange code.
