+++ unicorn/include/cocaine/detail/unicorn/zookeeper/create.hpp
@@ -15,11 +15,11 @@
 
 #pragma once
 
-#include "cocaine/unicorn/api.hpp"
+#include "cocaine/api/unicorn.hpp"
 
-#include "cocaine/unicorn/api/zookeeper.hpp"
+#include "cocaine/detail/unicorn/zookeeper.hpp"
 
-#include "cocaine/zookeeper/handler.hpp"
+#include "cocaine/detail/zookeeper/handler.hpp"
 
 namespace cocaine { namespace unicorn {
