+++ unicorn/include/cocaine/detail/unicorn/zookeeper/increment.hpp
@@ -15,11 +15,13 @@
 
 #pragma once
 
-#include "cocaine/unicorn/api.hpp"
+#include "cocaine/api/unicorn.hpp"
 
-#include "cocaine/unicorn/api/zookeeper.hpp"
+#include "cocaine/detail/unicorn/zookeeper.hpp"
 
-#include "cocaine/zookeeper/handler.hpp"
+#include "cocaine/detail/unicorn/zookeeper/create.hpp"
+
+#include "cocaine/detail/zookeeper/handler.hpp"
 
 namespace cocaine { namespace unicorn {
