+++ unicorn/include/cocaine/detail/unicorn/zookeeper/lock.hpp
@@ -15,11 +15,14 @@
 
 #pragma once
 
+#include "cocaine/api/unicorn.hpp"
+
+#include "cocaine/detail/unicorn/zookeeper.hpp"
+
 #include "cocaine/detail/unicorn/zookeeper/create.hpp"
 #include "cocaine/detail/unicorn/zookeeper/lock_state.hpp"
-#include "cocaine/unicorn/api.hpp"
-#include "cocaine/unicorn/api/zookeeper.hpp"
-#include "cocaine/zookeeper/handler.hpp"
+
+#include "cocaine/detail/zookeeper/handler.hpp"
 
 namespace cocaine { namespace unicorn {
