+++ unicorn/include/cocaine/detail/unicorn/zookeeper/put.hpp
@@ -15,10 +15,11 @@
 
 #pragma once
 
-#include "cocaine/unicorn/api.hpp"
-#include "cocaine/unicorn/api/zookeeper.hpp"
+#include "cocaine/api/unicorn.hpp"
 
-#include "cocaine/zookeeper/handler.hpp"
+#include "cocaine/detail/unicorn/zookeeper.hpp"
+
+#include "cocaine/detail/zookeeper/handler.hpp"
 
 namespace cocaine { namespace unicorn {
