+++ unicorn/include/cocaine/service/unicorn.hpp
@@ -16,7 +16,7 @@
 #pragma once
 
 #include "cocaine/idl/unicorn.hpp"
-#include "cocaine/unicorn/api.hpp"
+#include "cocaine/api/unicorn.hpp"
 
 #include <cocaine/api/service.hpp>
 #include <cocaine/rpc/dispatch.hpp>
