+++ unicorn/src/unicorn/value.cpp
@@ -23,28 +23,4 @@ versioned_value_t::versioned_value_t(value_t _value, version_t _version) :
     version(std::move(_version))
 {}
 
-zookeeper::value_t serialize(const value_t& val) {
-    msgpack::sbuffer buffer;
-    msgpack::packer<msgpack::sbuffer> packer(buffer);
-    cocaine::io::type_traits<cocaine::dynamic_t>::pack(packer, val);
-    return std::string(buffer.data(), buffer.size());
-}
-
-value_t unserialize(const zookeeper::value_t& val) {
-    msgpack::object obj;
-    std::unique_ptr<msgpack::zone> z(new msgpack::zone());
-
-    msgpack_unpack_return ret = msgpack_unpack(
-        val.c_str(), val.size(), nullptr, z.get(),
-        reinterpret_cast<msgpack_object*>(&obj)
-    );
-
-    //Only strict unparse.
-    if(static_cast<msgpack::unpack_return>(ret) != msgpack::UNPACK_SUCCESS) {
-        throw std::system_error(cocaine::error::unicorn_errors::INVALID_VALUE);
-    }
-    value_t target;
-    cocaine::io::type_traits<cocaine::dynamic_t>::unpack(obj, target);
-    return target;
-}
 }}
